fix(stella-tui): move transcript_source's tests to a sibling file
`main` is red on `file-size`: `crates/stella-tui/src/views/transcript_source.rs`
is 1540 lines, over the 1500-line ceiling, with no baseline entry — and the
baseline takes none, because the rule that stops the tree getting worse is
that a first-time crossing gets split.

Nobody's branch was wrong. #5236 added the model row's projection to a file
that had room for it at the time; the ceiling is a shared cell, so the tree
only crossed once both halves were in it. That is the composition hazard the
post-merge canary exists to catch, and it caught this one.

The seam is where the file already had one: two thirds of it was an inline
`#[cfg(test)] mod tests`, so the split follows the shape `model/tests.rs`,
`deck/tests.rs` and `driver/tests.rs` already take in this workspace. The move
is verbatim — all 27 tests, same assertions, de-indented one level with
`use super::*` doing the work the enclosing module used to do. Production code
is untouched, which is what makes this reviewable as a move rather than a
rewrite.

Leaves the production file at 525 lines and the tests at 1022, so the next
projection to land has room and does not have to pay for this one.

Closes #5256
